Add registry to register all needed classes with artifacts in nemo.lightning.io
