+++ nemo/lightning/io/__init__.py
@@ -1,10 +1,10 @@
 from nemo.lightning.io.api import export_ckpt, import_ckpt, load, load_context, model_exporter, model_importer
 from nemo.lightning.io.capture import reinit
 from nemo.lightning.io.connector import Connector, ModelConnector
-from nemo.lightning.io.mixin import ConnectorMixin, IOMixin, track_io
+from nemo.lightning.io.mixin import ConnectorMixin, IOMixin
 from nemo.lightning.io.pl import TrainerContext, is_distributed_ckpt
+from nemo.lightning.io.registry import track_io
 from nemo.lightning.io.state import TransformCTX, apply_transforms, state_transform
-
 
 __all__ = [
     "apply_transforms",
