+++ nemo/lightning/io/api.py
@@ -1,9 +1,7 @@
 from pathlib import Path
-from typing import Any, Callable, Optional, Type, TypeVar
+from typing import Callable, Optional, Type
 
-import fiddle as fdl
 import pytorch_lightning as pl
-from fiddle._src.experimental import serialization
 
 from nemo.lightning.io.mixin import ConnectorMixin, ConnT, ModelConnector, load
 from nemo.lightning.io.pl import TrainerContext
