+++ tests/collections/llm/test_mnist_model_nemo2_fsdp.py
@@ -39,6 +39,7 @@ from torch.utils.data import DataLoader
 from torchvision import transforms
 from torchvision.datasets import MNIST
 
+import nemo.lightning.io.registry
 from nemo import lightning as nl
 from nemo.collections import llm
 from nemo.lightning import NeMoLogger, io, resume
@@ -550,7 +551,7 @@ def run_train_mnist_litautoencoder_with_fsdp_strategy_single_gpu():
                 max_steps=20,
                 num_nodes=1,
                 log_every_n_steps=5,
-                callbacks=[io.track_io(LossLoggingCallback)()],
+                callbacks=[nemo.lightning.io.registry.track_io(LossLoggingCallback)()],
             )
             data_module = MNISTDataModule(data_dir=tmpdir)
             llm.train(
